BIG-IQ 5.0 support

+++ lib/f5_image_prep/os-functions/os-license_via_bigiq.sh
@@ -1,12 +1,39 @@
 #!/bin/bash
+#
+# Copyright 2015-2016 F5 Networks Inc.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+# 2015/12/15 - u.alonsocamaro@f5.com - First version released in hive
+# 2016/04/29 - u.alonsocamaro@f5.com - First version released in github supporting BIG-IQ 4.5/4.6
+# 2016/09/22 - u.alonsocamaro@f5.com - Support of BIG-IQ 5.0. Registration and registration are now in their own separate functions
 
 shopt -s extglob
 source /config/os-functions/os-functions.sh
 
-readonly OS_BIGIQ_LICENSE_POOL_USER=admin
-readonly OS_BIGIQ_LICENSE_POOL_PASSWORD=admin
+# Defaults
+
+readonly OS_BIGIQ_ADMIN_PASSWORD=admin
+readonly OS_BIGIQ_ROOT_PASSWORD=default
+
 readonly OS_BIGIQ_LICENSE_POOL_UUID=any
 readonly OS_BIGIQ_UPDATE_FRAMEWORK=true
+
+readonly OS_BIGIQ_LICENSE_POOL_HOST=127.0.0.1
+readonly OS_BIGIQ_LICENSE_POOL_USER=admin
+readonly OS_BIGIQ_LICENSE_POOL_PASSWORD=admin
+
+# 
 readonly OS_BIGIQ_JSON_REPLY_FILE=/tmp/bigiq_reply.json
 readonly OS_BIGIQ_JSON_REPLY_TMP_FILE=/tmp/bigiq_reply.tmp
 readonly OS_BIGIQ_MAX_RETRIES=20
@@ -74,22 +101,122 @@ function get_bigiq_reply_values_from_array ()
     echo -n $(get_json_values_from_array $1 $2 $OS_BIGIQ_JSON_REPLY_TMP_FILE)
 }
 
+# needs_bigiq_registration checks if the BIG-IQ version is less than 5.0.0
+
+function check_bigiq_5plus ()
+{
+
+    local http_code=$(curlbigiq "/mgmt/shared/resolver/device-groups/cm-shared-all-big-iqs/devices?\\\$select=version" -X GET )
+
+    local -a bqVersion
+
+    bqVersion=( $(get_bigiq_reply_values_from_array {items} {version}) )
+
+    [[ "${bqVersion[0]}" < "5.0.0" ]] && echo 0 || echo 1;
+}
+
+function register_bigip ()
+{
+        local bigip_mgmt_ip=$(get_mgmt_ip)
+        local JSON="{\"deviceAddress\": \"$bigip_mgmt_ip\", \"username\":\"admin\", \"password\":\"$bigip_admin_password\", \"automaticallyUpdateFramework\":\"$bigiq_update_framework\", \"rootUsername\":\"root\", \"rootPassword\":\"$bigip_root_password\"}"
+
+        local i
+	declare -i i
+	i=1
+        local state=""
+
+        while [[ -z "$state" ]]; do
+
+                http_code=$(curlbigiq /mgmt/cm/cloud/managed-devices -d "\"$JSON\"" -X POST)
+
+                local state=$(get_bigiq_reply_value {state})
+
+                if [ -z "$state" ] ; then
+                        if [[ $OS_BIGIQ_MAX_RETRIES -lt $i ]]; then
+                                log_bigiq_message "Error while registering this BIG-IP in BIG-IQ. Max retries reached ($i of $OS_BIGIQ_MAX_RETRIES). Aborting..."
+                                return 1
+                        else
+                                log_bigiq_message "Error while registering this BIG-IP in BIG-IQ. Retrying ($i of $OS_BIGIQ_MAX_RETRIES)..."
+                        fi
+                else
+                        log "BIG-IP registered in BIG-IQ. Waiting for ACTIVE state in BIG-IQ..."
+                        break;
+                fi
+
+                i=$i+1
+                sleep 10
+        done
+
+        local machineid=$(get_bigiq_reply_value {machineId})
+        local selflink=$(get_bigiq_reply_value {selfLink})
+
+        local code=""
+        local errors=""
+        local message=""
+
+        # Loop until we have an 'ACTIVE' state with the BIGIQ. When updating the framework this can take some time.
+        i=0
+
+        # Go get the BIGIQ's record for this BIGIP node
+        http_code=$(curlbigiq /mgmt/cm/cloud/managed-devices/$machineid -X GET)
+
+        state=$(get_bigiq_reply_value {state})
+
+        while [ "$state" != "ACTIVE" ];
+        do
+
+                if [ "$state" == "POST_FAILED" ]; then
+                        log_bigiq_errors "Error while waiting the BIG-IP to be active in the BIG-IQ"
+                        return 1
+                fi
+
+                if [ -z "$state" ]; then
+                        code=$(get_bigiq_reply_value {code})
+
+                        if [ "$code" = "404" ]; then
+                                log "Error the BIG-IP is not found registered in the BIG-IQ"
+                        fi
+                fi
+
+                if [ $i == $OS_BIGIQ_MAX_RETRIES ] && [ "$state" != "ACTIVE" ]; then
+                        log "Aborting licensing in the BIG-IQ too many retries while waiting for ACTIVE state"
+                        return 1
+                fi
+
+                i=$i+1
+                log "Waiting for ACTIVE state in BIG-IQ, current status: $state..."
+                sleep 5
+
+                # Go get the BIGIQ's record for this BIGIP node
+                http_code=$(curlbigiq $CBASE/mgmt/cm/cloud/managed-devices/$machineid -X GET)
+
+                state=$(get_bigiq_reply_value {state})
+        done
+
+
+	echo $selflink
+	return 0
+}
+
+function unregister_bigip() {
+
+	local http_code=$(curlbigiq /mgmt/cm/cloud/managed-devices/$OS_THISUUID -X DELETE )
+
+	if [ "$http_code" != "200" ]; then
+		log "Error while trying to delete this device with uuid $OS_THIS_UUIDi from BIG-IQ"
+		return 1
+	fi
+
+	log 'Could eliminate this device from BIG-IQ'
+	return 0
+}
+
 # licenese a BIG-IP using a BIG-IQ pool license. The pool license ID can be specified in the JSON file or
 # any can be specified in which case loops through all the licenses until it finds a valid one.
 #
 # Before the actual license the BIG-IP is registered in BIG-IQ
 
 function license_via_bigiq_license_pool() {
-
-	local bigip_admin_password=$1
-	local bigip_root_password=$2
-
-	local bigiq_license_pool_uuid=$(get_user_data_value {bigip}{license}{bigiq_license_pool_uuid})
-	local bigiq_update_framework=$(get_user_data_value {bigip}{license}{bigiq_update_framework})
-
-	local bigiq_license_pool_host=$(get_user_data_value {bigip}{license}{bigiq_license_pool_host})
-	local bigiq_license_pool_user=$(get_user_data_value {bigip}{license}{bigiq_license_pool_user})
-	local bigiq_license_pool_password=$(get_user_data_value {bigip}{license}{bigiq_license_pool_password})
 
 	local http_code
 	local JSON
@@ -101,88 +228,15 @@ function license_via_bigiq_license_pool() {
 		return 1
 	fi
 
-	[[ $(is_false ${bigiq_license_pool_user}) ]] && bigiq_license_pool_user=${OS_BIGIQ_LICENSE_POOL_USER}
-	[[ $(is_false ${bigiq_license_pool_password}) ]] && bigiq_license_pool_password=${OS_BIGIQ_LICENSE_POOL_PASSWORD}
-        [[ $(is_false ${bigiq_license_pool_uuid}) ]] && bigiq_license_pool_uuid=${OS_BIGIQ_LICENSE_POOL_UUID}
-	[[ $(is_false ${bigiq_update_framework}) ]] && bigiq_update_framework=${OS_BIGIQ_UPDATE_FRAMEWORK}
-	
+	is_bigiq_5plus=$( check_bigiq_5plus )
 
-	#-------------------[ Register the BIG-IP in BIG-IQ ]-------------------
+	if [[ "$is_bigiq_5plus" = 0 ]]; then
+		selflink=$( register_bigip )
+		if [[ $? -ne 0 ]]; then
 
-	bigip_mgmt_ip=$(get_mgmt_ip)
-	JSON="{\"deviceAddress\": \"$bigip_mgmt_ip\", \"username\":\"admin\", \"password\":\"$bigip_admin_password\", \"automaticallyUpdateFramework\":\"$bigiq_update_framework\", \"rootUsername\":\"root\", \"rootPassword\":\"$bigip_root_password\"}"
-
-	i=1
-	state=""
-
-	while [[ -z "$state" ]]; do
-
-		http_code=$(curlbigiq /mgmt/cm/cloud/managed-devices -d "\"$JSON\"" -X POST)
-
-	        local state=$(get_bigiq_reply_value {state})
-
-        	if [ -z "$state" ] ; then
-			if [[ $OS_BIGIQ_MAX_RETRIES -lt $i ]]; then
-                		log_bigiq_message "Error while registering this BIG-IP in BIG-IQ. Max retries reached ($i of $OS_BIGIQ_MAX_RETRIES). Aborting..."
-				return 1
-			else
-				log_bigiq_message "Error while registering this BIG-IP in BIG-IQ. Retrying ($i of $OS_BIGIQ_MAX_RETRIES)..."
-			fi
-		else
-			log "BIG-IP registered in BIG-IQ. Waiting for ACTIVE state in BIG-IQ..."
-			break;
+			return 1;
 		fi
-
-		i=$i+1
-		sleep 10
-	done
-	
-	local machineid=$(get_bigiq_reply_value {machineId})
-	local selflink=$(get_bigiq_reply_value {selfLink})
-
-	local code=""
-	local errors=""
-	local message=""
-	
-
-	# Loop until we have an 'ACTIVE' state with the BIGIQ. When updating the framework this can take some time.
-	i=0
-
-	# Go get the BIGIQ's record for this BIGIP node
-	http_code=$(curlbigiq /mgmt/cm/cloud/managed-devices/$machineid -X GET)
-
-        state=$(get_bigiq_reply_value {state})
-
-	while [ "$state" != "ACTIVE" ];
-	do
-	
-		if [ "$state" == "POST_FAILED" ]; then
-			log_bigiq_errors "Error while waiting the BIG-IP to be active in the BIG-IQ"
-			return 1
-		fi
-
-		if [ -z "$state" ]; then
-			code=$(get_bigiq_reply_value {code})
-			
-			if [ "$code" = "404" ]; then
-				log "Error the BIG-IP is not found registered in the BIG-IQ"
-			fi
-		fi
-	
-		if [ $i == $OS_BIGIQ_MAX_RETRIES ] && [ "$state" != "ACTIVE" ]; then
-			log "Aborting licensing in the BIG-IQ too many retries while waiting for ACTIVE state"
-			return 1	
-		fi
-	
-		i=$i+1
-		log "Waiting for ACTIVE state in BIG-IQ, current status: $state..."
-		sleep 5 
-
-                # Go get the BIGIQ's record for this BIGIP node
-                http_code=$(curlbigiq $CBASE/mgmt/cm/cloud/managed-devices/$machineid -X GET)
-
-                state=$(get_bigiq_reply_value {state})
-	done
+	fi
 
         #-------------------[ License the BIG-IP from the license pools in BIG-IQ ]-------------------
 
@@ -199,12 +253,20 @@ function license_via_bigiq_license_pool() {
 	# Try each license pool until we get one that works.
 	for pool in $pools; do
 
-		log "Trying to obtain a license from BIG-IQ's pool license $pool ..."
+		log "Trying to obtain a license from BIG-IP pool license $pool ..."
 
-                JSON='{\"deviceReference\":{\"link\": \"$selflink\"}}'
+        	if [[ "$is_bigiq_5plus" = 0 ]]; then
+
+                	JSON='{\"deviceReference\":{\"link\": \"$selflink\"}}'
+		else
+
+		        local bigip_mgmt_ip=$(get_mgmt_ip)
+
+			JSON="{\"deviceAddress\": \"$bigip_mgmt_ip\", \"username\": \"admin\", \"password\": \"$bigip_admin_password\"}"
+		fi
+
 		http_code=$(curlbigiq /mgmt/cm/shared/licensing/pools/$pool/members -X POST -d "\"$JSON\"")
 		# cp $OS_BIGIQ_JSON_REPLY_FILE $OS_BIGIQ_JSON_REPLY_FILE.post
-
 
                 uuid=$(get_bigiq_reply_value {uuid})
 
@@ -220,14 +282,6 @@ function license_via_bigiq_license_pool() {
 		while [ "$state" != "LICENSED" ];
 		do
 
-                        log "Waiting for LICENSED status in BIG-IQ, current status: $state..."
-			sleep 5
-
-			http_code=$(curlbigiq /mgmt/cm/shared/licensing/pools/$pool/members/$UUID -X GET)
-			# cp $OS_BIGIQ_JSON_REPLY_FILE $OS_BIGIQ_JSON_REPLY_FILE.get
-
-	        	state=$(get_bigiq_reply_value {items}[0]{state})
-			
 			if [ -z "$state" ]; then
 
 	                        code=$(get_bigiq_reply_value {items}[0]{code})
@@ -241,13 +295,22 @@ function license_via_bigiq_license_pool() {
 				log "Error while licensing the BIG-IP with pool license $uuid: no state has been returned"
 				return 1
 			fi
-		
+	
+                        i=$i+1
+	
 			if [ $i == $OS_BIGIQ_MAX_RETRIES ] && [ "$state" != "LICENSED" ]; then
 	                        log "Aborting licensing in the BIG-IQ too many retries while waiting for LICENSED state"
 				return 1
 			fi
 
-			i=$i+1
+                        log "Waiting for LICENSED status in BIG-IQ, current status: $state..."
+                        sleep 5
+
+                        http_code=$(curlbigiq /mgmt/cm/shared/licensing/pools/$pool/members/$UUID -X GET)
+                        # cp $OS_BIGIQ_JSON_REPLY_FILE $OS_BIGIQ_JSON_REPLY_FILE.get
+
+                        state=$(get_bigiq_reply_value {items}[0]{state})
+
 		done
 
 		local expired=$(tmsh show sys license | grep "^Warning" | cut -d' ' -f4)
@@ -268,6 +331,8 @@ function license_via_bigiq_license_pool() {
 		
 	done
 
+	log "Licensing failure: didn't find any pool in BIG-IQ"
+
 	return 1
 }
 
@@ -278,35 +343,49 @@ function license_via_bigiq_license_pool() {
 
 function unlicense_via_bigiq_license_pool() {
 
-	local bigiq_license_pool_host=$(get_user_data_value {bigip}{license}{bigiq_license_pool_host})
-	local bigiq_license_pool_user=$(get_user_data_value {bigip}{license}{bigiq_license_pool_user})
-	local bigiq_license_pool_password=$(get_user_data_value {bigip}{license}{bigiq_license_pool_password})
+	local http_code
+	local -a pools
+	local memberIds
+	local -a uuids
+	local -i i
 
 	# Retrieve all the license Pools on BIGIQ
 	http_code=$(curlbigiq /mgmt/cm/shared/licensing/pools/?\$select=uuid -X GET)
-	# cp $OS_BIGIQ_JSON_REPLY_FILE $OS_BIGIQ_JSON_REPLY_FILE.pools
-	pools=$(get_bigiq_reply_values_from_array {items} {uuid})
+
+	pools=( $(get_bigiq_reply_values_from_array {items} {uuid}) )
+
+	is_bigiq_5plus=$( check_bigiq_5plus )
 
         # Try each license pool until we find us 
         for pool in $pools; do
 
 		http_code=$(curlbigiq /mgmt/cm/shared/licensing/pools/$pool/members)
-		# cp $OS_BIGIQ_JSON_REPLY_FILE $OS_BIGIQ_JSON_REPLY_FILE.members
-		local devices=$( get_bigiq_reply_values_from_array {items} {deviceReference}{link} )
 
-		local -i i
-		local -a uuids
+		# Let's make version independent Ids
+		if [[ "$is_bigiq_5plus" = 1 ]]; then
+			memberIds=$( get_bigiq_reply_values_from_array {items} {deviceMachineId} )
+			myId="$OS_THISUUID"
+		else
+			memberIds=$( get_bigiq_reply_values_from_array {items} {deviceReference}{link} )
+			myId="$OS_THISDEVICE"
+		fi
 
 		uuids=($( get_bigiq_reply_values_from_array {items} {uuid} ))
-     
+
 		i=0 
-		for device in $devices; do
- 
-                	if [ "$device" = "$OS_THISDEVICE" ]; then
+		for memberId in $memberIds; do
+
+			if [[ "$memberId" = "$myId" ]]; then
 
                         	uuid=${uuids[$i]}
 
-				http_code=$(curlbigiq /mgmt/cm/shared/licensing/pools/$pool/members/$uuid -X DELETE )
+		                if [[ "$is_bigiq_5plus" = 1 ]]; then
+
+	                        	JSON="{\"uuid\": \"$uuid\", \"username\": \"admin\", \"password\": \"$bigip_admin_password\"}"
+					http_code=$(curlbigiq /mgmt/cm/shared/licensing/pools/$pool/members/$uuid -X DELETE -d "\"$JSON\"" )
+				else
+					http_code=$(curlbigiq /mgmt/cm/shared/licensing/pools/$pool/members/$uuid -X DELETE )
+				fi	
 
 				if [ "$http_code" != "200" ]; then
                                 	log "Error while trying to release the license $uuid in pool $pool"
@@ -315,14 +394,13 @@ function unlicense_via_bigiq_license_pool() {
 
 				log "Could eliminate license $uuid from license pool $pool"
 
-				http_code=$(curlbigiq /mgmt/cm/cloud/managed-devices/$OS_THISUUID -X DELETE )
+			        is_bigiq_5plus=$( check_bigiq_5plus )
 
-				if [ "$http_code" != "200" ]; then
-					log "Error while trying to delete this device with uuid $OS_THIS_UUIDi from BIG-IQ"
-					return 1
-				fi
+			        if [[ "$is_bigiq_5plus" = 0 ]]; then
+					unregister_bigip
+                			return $?
+        			fi
 
-				log 'Could eliminate this device from BIG-IQ'
 				return 0
 			fi
 
@@ -341,6 +419,26 @@ function unlicense_via_bigiq_license_pool() {
 	return 1
 }
 
+function do_license_via_bigiq() {
+
+        license_via_bigiq_license_pool
+        if [[ $? = 0 ]]; then
+                echo "license_via_bigiq_license_pool succeeded"
+        else
+                echo "license_via_bigiq_license_pool failed"
+        fi
+}
+
+function do_unlicense_via_bigiq() {
+
+        unlicense_via_bigiq_license_pool
+        if [[ $? = 0 ]]; then
+                echo "unlicense_via_bigiq_license_pool succeeded"
+        else
+                echo "unlicense_via_bigiq_license_pool failed"
+        fi
+}
+
 function test() {
 
 	rm -f /config/bigip.license
@@ -350,38 +448,59 @@ function test() {
 
 	set -x
 
-	license_via_bigiq_license_pool admin default        
-        if [[ $? = 0 ]]; then
-                echo "license_via_bigiq_license_pool succeeded"
-        else
-                echo "license_via_bigiq_license_pool failed"
-        fi
+	do_license_via_bigiq
 
 	sleep 5
 
-	unlicense_via_bigiq_license_pool 
-        if [[ $? = 0 ]]; then
-                echo "unlicense_via_bigiq_license_pool succeeded"
-        else
-                echo "unlicense_via_bigiq_license_pool failed"
-        fi
+	do_unlicense_via_bigiq
 
 	set +x
 }
 
+
+### MAIN #########################################################
+
+get_user_data
+
+bigip_admin_password=$(get_user_data_value {bigip}{admin_password})
+bigip_root_password=$(get_user_data_value {bigip}{root_password})
+
+bigiq_license_pool_uuid=$(get_user_data_value {bigip}{license}{bigiq_license_pool_uuid})
+bigiq_update_framework=$(get_user_data_value {bigip}{license}{bigiq_update_framework})
+
+bigiq_license_pool_host=$(get_user_data_value {bigip}{license}{bigiq_license_pool_host})
+bigiq_license_pool_user=$(get_user_data_value {bigip}{license}{bigiq_license_pool_user})
+bigiq_license_pool_password=$(get_user_data_value {bigip}{license}{bigiq_license_pool_password})
+
+
+[[ $(is_false ${bigiq_admin_password}) ]] && bigiq_admin_password=${OS_BIGIQ_ADMIN_PASSWORD}
+[[ $(is_false ${bigiq_root_password}) ]] && bigiq_root_password=${OS_BIGIQ_ROOT_PASSWORD}
+
+[[ $(is_false ${bigiq_license_pool_uuid}) ]] && bigiq_license_pool_uuid=${OS_BIGIQ_LICENSE_POOL_UUID}
+[[ $(is_false ${bigiq_update_framework}) ]] && bigiq_update_framework=${OS_BIGIQ_UPDATE_FRAMEWORK}
+
+[[ $(is_false ${bigiq_license_pool_host}) ]] && bigiq_license_pool_host=${OS_BIGIQ_LICENSE_POOL_HOST}
+[[ $(is_false ${bigiq_license_pool_user}) ]] && bigiq_license_pool_user=${OS_BIGIQ_LICENSE_POOL_USER}
+[[ $(is_false ${bigiq_license_pool_password}) ]] && bigiq_license_pool_password=${OS_BIGIQ_LICENSE_POOL_PASSWORD}
+
 if [[ $1 = "test" ]]; then
 	test
+elif [[ $1 = "license" ]]; then
+
+        do_license_via_bigiq
 elif [[ $1 = "unlicense" ]]; then
 
-        get_user_data
-        unlicense_via_bigiq_license_pool
-        if [[ $? = 0 ]]; then
-                echo "unlicense_via_bigiq_license_pool succeeded"
-		exit 0
-        else
-                echo "unlicense_via_bigiq_license_pool failed"
-		exit 1
-        fi
+	do_unlicense_via_bigiq
+elif [[ $1 = "register" ]]; then
+
+	echo Registered device with selfLink...
+	register_bigip
+
+elif [[ $1 = "unregister" ]]; then
+
+	unregister_bigip
+else
+	echo "Nothing to do: check script's parameters"
 fi
